rag/requirements.txt
pymongo
pinecone-client
python-dotenv
fastapi
uvicorn